cpuquiet: revert runnable code
note to self;

Currently the system is utilizing the balanced governor, but even this is not ideal as
cpuquiet is a very archaic hot plugging driver. There seems to be no proper documentation on
the cpuquiet driver.

Stuttering is inevitable with newer android systems, but disabling it entirely is not a proper
solution because it also initiates a proper deep sleep mode.

Cpuquiet for the future should be placed in user-space, all cores will online when the screen
is on, and 3 cores will offline when the screen is off. This eliminates stuttering and saves
battery power since there are no constant hot plugging calculations going on.